rename: MimeType 클래스명 변경
